Fix up size

+++ stdlib/stdint.h
@@ -2,5 +2,5 @@
 
 typedef unsigned long long uint64_t;
 typedef unsigned char uint8_t;
-typedef unsigned int uint32_t;
+typedef unsigned long uint32_t;
 typedef uint32_t size_t;
